[Fix] modify git gitconfig.txt and set_git.sh
1. remove the user information from gitconfig.txt
2. add some if statment for check the alias label in ~/.gitconfig

+++ set_git.sh
@@ -1,2 +1,12 @@
-[ -f ~/.gitconfig ] && cp ~/.gitconfig ~/.gitconfig.backup
-cp gitconfig.txt ~/.gitconfig
+if [ -f ~/.gitconfig ];then
+       	cp ~/.gitconfig ~/.gitconfig.backup
+        if grep alias ~/.gitconfig ;then
+		echo 'Pass'
+		continue
+	else
+		cat gitconfig.txt >> ~/.gitconfig
+		echo 'Add alias'
+	fi
+else
+	cp gitconfig.txt ~/.gitconfig
+fi
